+++ tests/integration_tests/functional/test_logging.py
@@ -5,11 +5,12 @@
 It checks the response of the API configuration calls and the logs that show
 up in the configured logging FIFO.
 """
-import os
+
 import re
+from pathlib import Path
 from time import strptime
 
-import host_tools.logging as log_tools
+import pytest
 
 # Array of supported log levels of the current logging system.
 # Do not change order of values inside this array as logic depends on this.
@@ -117,19 +118,20 @@ def test_log_config_failure(test_microvm_with_api):
     Check passing invalid FIFOs is detected and reported as an error.
     """
     microvm = test_microvm_with_api
-    microvm.spawn(create_logger=False)
+    microvm.spawn(log_file=None)
     microvm.basic_config()
 
-    response = microvm.logger.put(
-        log_path="invalid log fifo",
-        level="Info",
-        show_level=True,
-        show_log_origin=True,
-    )
     # only works if log level is Debug
     microvm.time_api_requests = False
-    assert microvm.api_session.is_status_bad_request(response.status_code)
-    assert response.json()["fault_message"]
+
+    expected_msg = re.escape("No such file or directory (os error 2)")
+    with pytest.raises(RuntimeError, match=expected_msg):
+        microvm.api.logger.put(
+            log_path="invalid log file",
+            level="Info",
+            show_level=True,
+            show_log_origin=True,
+        )
 
 
 def test_api_requests_logs(test_microvm_with_api):
@@ -137,27 +139,24 @@ def test_api_requests_logs(test_microvm_with_api):
     Test that API requests are logged.
     """
     microvm = test_microvm_with_api
-    microvm.spawn(create_logger=False)
+    microvm.spawn(log_file=None)
     microvm.basic_config()
 
     # Configure logging.
-    log_fifo_path = os.path.join(microvm.path, "log_fifo")
-    log_fifo = log_tools.Fifo(log_fifo_path)
-
-    response = microvm.logger.put(
-        log_path=microvm.create_jailed_resource(log_fifo.path),
+    log_path = Path(microvm.path) / "log"
+    log_path.touch()
+    microvm.api.logger.put(
+        log_path=microvm.create_jailed_resource(log_path),
         level="Info",
         show_level=True,
         show_log_origin=True,
     )
+    microvm.log_file = log_path
     # only works if log level is Debug
     microvm.time_api_requests = False
-    assert microvm.api_session.is_status_no_content(response.status_code)
-    microvm.start_console_logger(log_fifo)
 
     # Check that a Patch request on /machine-config is logged.
-    response = microvm.machine_cfg.patch(vcpu_count=4)
-    assert microvm.api_session.is_status_no_content(response.status_code)
+    microvm.api.machine_config.patch(vcpu_count=4)
     # We are not interested in the actual body. Just check that the log
     # message also has the string "body" in it.
     microvm.check_log_message(
@@ -165,68 +164,60 @@ def test_api_requests_logs(test_microvm_with_api):
     )
 
     # Check that a Put request on /machine-config is logged.
-    response = microvm.machine_cfg.put(vcpu_count=4, mem_size_mib=128)
-    assert microvm.api_session.is_status_no_content(response.status_code)
+    microvm.api.machine_config.put(vcpu_count=4, mem_size_mib=128)
     microvm.check_log_message(
         "The API server received a Put request " 'on "/machine-config" with body'
     )
 
     # Check that a Get request on /machine-config is logged without the
     # body.
-    response = microvm.machine_cfg.get()
-    assert microvm.api_session.is_status_ok(response.status_code)
+    microvm.api.machine_config.get()
     microvm.check_log_message(
         "The API server received a Get request " 'on "/machine-config".'
     )
 
     # Check that all requests on /mmds are logged without the body.
     dummy_json = {"latest": {"meta-data": {"ami-id": "dummy"}}}
-    response = microvm.mmds.put(json=dummy_json)
-    assert microvm.api_session.is_status_no_content(response.status_code)
+    microvm.api.mmds.put(json=dummy_json)
     microvm.check_log_message('The API server received a Put request on "/mmds".')
 
-    response = microvm.mmds.patch(json=dummy_json)
-    assert microvm.api_session.is_status_no_content(response.status_code)
+    microvm.api.mmds.patch(json=dummy_json)
     microvm.check_log_message('The API server received a Patch request on "/mmds".')
 
-    response = microvm.mmds.get()
-    assert microvm.api_session.is_status_ok(response.status_code)
+    microvm.api.mmds.get()
     microvm.check_log_message('The API server received a Get request on "/mmds".')
 
     # Check that the fault message return by the client is also logged in the
     # FIFO.
-    response = microvm.boot.put(kernel_image_path="inexistent_path")
-    assert microvm.api_session.is_status_bad_request(response.status_code)
-    fault_message = (
-        "The kernel file cannot be opened: No such file or " "directory (os error 2)"
+    fault_msg = (
+        "The kernel file cannot be opened: No such file or directory (os error 2)"
     )
-    assert fault_message in response.text
+    with pytest.raises(RuntimeError, match=re.escape(fault_msg)):
+        microvm.api.boot.put(kernel_image_path="inexistent_path")
     microvm.check_log_message(
         "Received Error. "
         "Status code: 400 Bad Request. "
-        "Message: {}".format(fault_message)
+        "Message: {}".format(fault_msg)
     )
 
 
 # pylint: disable=W0102
 def _test_log_config(microvm, log_level="Info", show_level=True, show_origin=True):
     """Exercises different scenarios for testing the logging config."""
-    microvm.spawn(create_logger=False)
+    microvm.spawn(log_file=None)
     # only works if log level is Debug
     microvm.time_api_requests = False
 
     # Configure logging.
-    log_fifo_path = os.path.join(microvm.path, "log_fifo")
-    log_fifo = log_tools.Fifo(log_fifo_path)
-    response = microvm.logger.put(
-        log_path=microvm.create_jailed_resource(log_fifo.path),
+    log_path = Path(microvm.path) / "log"
+    log_path.touch()
+    microvm.api.logger.put(
+        log_path=microvm.create_jailed_resource(log_path),
         level=log_level,
         show_level=show_level,
         show_log_origin=show_origin,
     )
-    assert microvm.api_session.is_status_no_content(response.status_code)
-
-    microvm.start_console_logger(log_fifo)
+    microvm.log_file = log_path
 
     microvm.basic_config()
     microvm.start()