+++ tests/integration_tests/functional/test_max_devices.py
@@ -57,10 +57,9 @@ def test_attach_too_many_devices(test_microvm_with_api):
 
     # Attempting to start a microVM with more than
     # `MAX_DEVICES_ATTACHED` devices should fail.
-    response = test_microvm.actions.put(action_type="InstanceStart")
-    assert test_microvm.api_session.is_status_bad_request(response.status_code)
     error_str = (
         "Failed to allocate requested resource: The requested resource"
         " is not available."
     )
-    assert error_str in response.text
+    with pytest.raises(RuntimeError, match=error_str):
+        test_microvm.start()