+++ tests/integration_tests/functional/test_snapshot_version.py
@@ -51,21 +51,19 @@ def test_create_with_too_many_devices(test_microvm_with_api):
     _create_and_start_microvm_with_net_devices(test_microvm, devices_no)
 
     # Pause microVM for snapshot.
-    response = test_microvm.vm.patch(state="Paused")
-    assert test_microvm.api_session.is_status_no_content(response.status_code)
+    test_microvm.pause()
 
     # Attempt to create a snapshot with version: `0.23.0`. Firecracker
     # v0.23 allowed a maximum of `FC_V0_23_MAX_DEVICES_ATTACHED` virtio
     # devices at a time. This microVM has `FC_V0_23_MAX_DEVICES_ATTACHED`
     # network devices on top of the rootfs, so the limit is exceeded.
-    response = test_microvm.snapshot.create(
-        mem_file_path="/vm.mem",
-        snapshot_path="/vm.vmstate",
-        diff=True,
-        version="0.23.0",
-    )
-    assert test_microvm.api_session.is_status_bad_request(response.status_code)
-    assert "Too many devices attached" in response.text
+    with pytest.raises(RuntimeError, match="Too many devices attached"):
+        test_microvm.api.snapshot_create.put(
+            mem_file_path="/vm.mem",
+            snapshot_path="/vm.vmstate",
+            snapshot_type="Diff",
+            version="0.23.0",
+        )
 
 
 def test_create_invalid_version(uvm_nano):
@@ -76,33 +74,25 @@ def test_create_invalid_version(uvm_nano):
     test_microvm = uvm_nano
     test_microvm.start()
 
-    try:
-        # Target an invalid Firecracker version string.
-        test_microvm.pause_to_snapshot(
+    # Target an invalid Firecracker version string.
+    with pytest.raises(RuntimeError, match="unexpected character 'i'"):
+        test_microvm.api.snapshot_create.put(
             mem_file_path="/vm.mem",
             snapshot_path="/vm.vmstate",
-            diff=False,
+            snapshot_type="Full",
             version="invalid",
         )
-    except AssertionError as error:
-        # Check if proper error is returned.
-        assert "unexpected character 'i' while parsing major version" in str(error)
-    else:
-        assert False, "Negative test failed"
 
-    try:
-        # Target a valid version string but with no snapshot support.
-        test_microvm.pause_to_snapshot(
+    # Target a valid version string but with no snapshot support.
+    with pytest.raises(
+        RuntimeError, match="Cannot translate microVM version to snapshot data version"
+    ):
+        test_microvm.api.snapshot_create.put(
             mem_file_path="/vm.mem",
             snapshot_path="/vm.vmstate",
-            diff=False,
+            snapshot_type="Full",
             version="0.22.0",
         )
-    except AssertionError as error:
-        # Check if proper error is returned.
-        assert "Cannot translate microVM version to snapshot data version" in str(error)
-    else:
-        assert False, "Negative test failed"
 
 
 def test_snapshot_current_version(uvm_nano):
@@ -146,8 +136,7 @@ def test_create_with_newer_virtio_features(uvm_nano):
     test_microvm.ssh.run("true")
 
     # Pause microVM for snapshot.
-    response = test_microvm.vm.patch(state="Paused")
-    assert test_microvm.api_session.is_status_no_content(response.status_code)
+    test_microvm.pause()
 
     # We try to create a snapshot to a target version < 1.0.0.
     # This should fail because Fc versions < 1.0.0 don't support
@@ -155,39 +144,34 @@ def test_create_with_newer_virtio_features(uvm_nano):
     target_fc_versions = ["0.24.0", "0.25.0"]
     if platform.machine() == "x86_64":
         target_fc_versions.insert(0, "0.23.0")
+
+    expected_msg = (
+        "The virtio devices use a features that is incompatible "
+        "with older versions of Firecracker: notification suppression"
+    )
     for target_fc_version in target_fc_versions:
-        response = test_microvm.snapshot.create(
-            mem_file_path="/vm.mem",
-            snapshot_path="/vm.vmstate",
-            version=target_fc_version,
-        )
-        assert test_microvm.api_session.is_status_bad_request(response.status_code)
-        assert (
-            "The virtio devices use a features that is incompatible "
-            "with older versions of Firecracker: notification suppression"
-            in response.text
-        )
+        with pytest.raises(RuntimeError, match=expected_msg):
+            test_microvm.api.snapshot_create.put(
+                mem_file_path="/vm.mem",
+                snapshot_path="/vm.vmstate",
+                version=target_fc_version,
+            )
 
     # We try to create a snapshot for target version 1.0.0. This should
     # fail because in 1.0.0 we do not support notification suppression for Net.
-    response = test_microvm.snapshot.create(
-        mem_file_path="/vm.mem",
-        snapshot_path="/vm.vmstate",
-        version="1.0.0",
-    )
-    assert test_microvm.api_session.is_status_bad_request(response.status_code)
-    assert (
-        "The virtio devices use a features that is incompatible "
-        "with older versions of Firecracker: notification suppression" in response.text
-    )
+    with pytest.raises(RuntimeError, match=expected_msg):
+        test_microvm.api.snapshot_create.put(
+            mem_file_path="/vm.mem",
+            snapshot_path="/vm.vmstate",
+            version="1.0.0",
+        )
 
     # It should work when we target a version >= 1.1.0
-    response = test_microvm.snapshot.create(
+    test_microvm.api.snapshot_create.put(
         mem_file_path="/vm.mem",
         snapshot_path="/vm.vmstate",
         version="1.1.0",
     )
-    assert test_microvm.api_session.is_status_no_content(response.status_code)
 
 
 def test_create_with_1_5_cpu_template(uvm_plain):
@@ -198,28 +182,19 @@ def test_create_with_1_5_cpu_template(uvm_plain):
     """
 
     # We remove KVM_CAP_IOEVENTFD from kvm checks just for testing purpose.
-    custom_cpu_template = json.loads("""{"kvm_capabilities": ["!36"]}""")
+    custom_cpu_template = json.loads('{"kvm_capabilities": ["!36"]}')
 
     test_microvm = uvm_plain
     test_microvm.spawn()
     test_microvm.basic_config(vcpu_count=2, mem_size_mib=256)
-    test_microvm.cpu_config(custom_cpu_template)
+    test_microvm.api.cpu_config.put(**custom_cpu_template)
     test_microvm.start()
 
-    test_microvm.pause()
     # Should fail because target version is less than 1.5
-    response = test_microvm.snapshot.create(
-        mem_file_path="/vm.mem",
-        snapshot_path="/vm.vmstate",
-        version="1.4.0",
-    )
-    assert test_microvm.api_session.is_status_bad_request(response.status_code)
-    assert "Cannot translate microVM version to snapshot data version" in response.text
+    with pytest.raises(
+        RuntimeError, match="Cannot translate microVM version to snapshot data version"
+    ):
+        test_microvm.snapshot_full(target_version="1.4.0")
 
-    # Should pass because target version is 1.5
-    response = test_microvm.snapshot.create(
-        mem_file_path="/vm.mem",
-        snapshot_path="/vm.vmstate",
-        version="1.5.0",
-    )
-    assert test_microvm.api_session.is_status_no_content(response.status_code)
+    # Should pass because target version is >=1.5
+    test_microvm.snapshot_full()